# graph/db+sqldb: different defaults for SQLite and Postgres query options