--- a/config_builder.go
+++ b/config_builder.go
@@ -1130,9 +1130,8 @@ func (d *DefaultDatabaseBuilder) BuildDatabase(
 					continue
 				}
 
-				migFn, ok := getSQLMigration(
+				migFn, ok := d.getSQLMigration(
 					ctx, version, dbs.ChanStateDB.Backend,
-					*d.cfg.ActiveNetParams.GenesisHash,
 				)
 				if !ok {
 					continue