--- a/sample-lnd.conf
+++ b/sample-lnd.conf
@@ -1601,6 +1601,12 @@
 ; Whether to skip executing schema migrations.
 ; db.postgres.skipmigrations=false
 
+; The maximum number of elements to use in a native-SQL batch query IN clause.
+; db.postgres.query.max-batch-size=5000
+
+; The maximum number of records to fetch at a time in a paginated query during
+; native-SQL calls.
+; db.postgres.query.max-page-size=10500
 
 [sqlite]
 
@@ -1616,6 +1622,13 @@
 
 ; The maximum amount of time to wait to execute a query if the db is locked.
 ; db.sqlite.busytimeout=5s
+
+; The maximum number of elements to use in a native-SQL batch query IN clause.
+; db.sqlite.query.max-batch-size=250
+
+; The maximum number of records to fetch at a time in a paginated query during
+; native-SQL calls.
+; db.sqlite.query.max-page-size=100
 
 ; Raw pragma option pairs to be used when opening the sqlite db. The flag
 ; can be specified multiple times to set multiple options.