This PR does 2 things related to the `sqldb.QueryConfig`: 

1) plugs these values in to the main SQLite and Postgres config structs & makes the values configurable 
2) adds a benchmark which aided in finding better default values to use for sqlite and postgres for the query options. 

Part of #9795 

#  Benchmark improvements: 

### SQLite
```
name                                                    old time/op  new time/op  delta
GraphReadMethods/ForEachNode-native-sqlite-10            269ms ± 1%   200ms ± 3%  -25.86%  (p=0.008 n=5+5)
GraphReadMethods/ForEachChannel-native-sqlite-10         1.53s ±12%   1.40s ± 1%   -8.58%  (p=0.016 n=5+4)
GraphReadMethods/NodeUpdatesInHorizon-native-sqlite-10   238ms ± 2%   239ms ± 4%     ~     (p=0.841 n=5+5)
GraphReadMethods/ForEachNodeCacheable-native-sqlite-10   136ms ± 1%   114ms ± 5%  -15.92%  (p=0.008 n=5+5)
GraphReadMethods/ForEachNodeCached-native-sqlite-10      10.5s ± 5%    7.2s ± 1%  -31.26%  (p=0.008 n=5+5)
GraphReadMethods/ChanUpdatesInHorizon-native-sqlite-10   1.99s ± 3%   1.98s ± 4%     ~     (p=1.000 n=5+5)
```

### Postgres
```
name                                                      old time/op  new time/op  delta
GraphReadMethods/ForEachNode-native-postgres-10           91.1ms ± 1%  91.9ms ± 1%     ~     (p=0.111 n=5+4)
GraphReadMethods/ForEachChannel-native-postgres-10         601ms ± 5%   419ms ± 2%  -30.24%  (p=0.008 n=5+5)
GraphReadMethods/NodeUpdatesInHorizon-native-postgres-10  81.1ms ± 3%  79.5ms ± 5%     ~     (p=0.222 n=5+5)
GraphReadMethods/ForEachNodeCacheable-native-postgres-10  59.4ms ±10%  51.1ms ± 5%  -14.07%  (p=0.008 n=5+5)
GraphReadMethods/ForEachNodeCached-native-postgres-10      1.13s ± 4%   0.88s ± 2%  -22.10%  (p=0.008 n=5+5)
GraphReadMethods/ChanUpdatesInHorizon-native-postgres-10   784ms ± 3%   629ms ± 2%  -19.78%  (p=0.008 n=5+5)
```