@ellemouton 2025-08-13
@guggero  - thanks for the review :) 

i think the itest failure is not related as it is invoices related which does not make use of these query values yet.